Add files via upload
These files are the mod files from/for the Yuri Route, done by Jan. As of now (7/26/21), these files are untouched, and untested. I and the rest of the Yuri team will continue to edit and polish these up as we can (as there were stated a couple things to be added, among proofreading and such). We'll try to provide as much commentary on the changes as possible on the files, if things change.